+++ packages/nextjs-kit/src/components/contentWithImage.tsx
@@ -11,6 +11,7 @@ interface ContentProps {
 		src: ImageProps['src'];
 		alt?: ImageProps['alt'];
 	};
+	contentClassName?: string;
 }
 
 /**
@@ -20,6 +21,7 @@ interface ContentProps {
  * @param props.content - Content from a CMS, usually a string of raw HTML. This string will be set as `dangerouslySetInnerHTML`
  * @param props.date - An optional date to be displayed on the post
  * @param props.imageProps - Accepts a src and optional alt text for the next/image component. @see {@link https://nextjs.org/docs/api-reference/next/image} for more information.
+ * @param props.contentClassName - An optional class name to be applied to the content container
  * @remarks
  * `imageProps` is an optional prop to be used if there is an image associated with the content.
  * If `imageProps.src` is a supplied as a prop. Alt text is not required; however,
@@ -32,20 +34,23 @@ const ContentWithImage: React.FC<ContentProps> = ({
 	content,
 	date,
 	imageProps,
+	contentClassName = 'ps-max-w-screen lg:ps-max-w-screen-lg md:ps-max-w-screen-md sm:ps-max-w-screen-sm ps-mx-auto',
 }: ContentProps) => {
 	const router = useRouter();
 
 	return (
-		<article className="ps-prose xs:ps-prose-xs md:ps-prose-md lg:ps-prose-lg ps-mt-10 ps-mx-auto ps-max-w-screen lg:ps-max-w-screen-lg md:ps-max-w-screen-md sm:ps-max-w-screen-sm ps-p-4">
-			<h1>{title}</h1>
-			{date ? <p className="ps-text-sm ps-text-gray-600">{date}</p> : null}
+		<article className="ps-prose ps-max-w-none xs:ps-prose-xs md:ps-prose-md lg:ps-prose-lg ps-mt-10 ps-mx-auto ps-py-4 ps-px-12">
+			<section className="ps-prose xs:ps-prose-xs md:ps-prose-md lg:ps-prose-lg ps-mt-10 ps-max-w-screen lg:ps-max-w-screen-lg md:ps-max-w-screen-md sm:ps-max-w-screen-sm ps-mx-auto">
+				<h1>{title}</h1>
+				{date ? <p className="ps-text-sm ps-text-gray-600">{date}</p> : null}
 
-			<a
-				onClick={() => router.back()}
-				className="ps-font-normal ps-cursor-pointer"
-			>
-				Back &rarr;
-			</a>
+				<a
+					onClick={() => router.back()}
+					className="ps-font-normal ps-cursor-pointer"
+				>
+					Back &rarr;
+				</a>
+			</section>
 			<div className="ps-mt-12 ps-max-w-screen ps-mx-auto lg:ps-max-w-screen-lg ps-shadow-lg [&*>img]:ps-rounded-lg">
 				{imageProps ? (
 					<div className="ps-relative ps-mb-10 ps-min-h-[50vh]">
@@ -61,7 +66,7 @@ const ContentWithImage: React.FC<ContentProps> = ({
 			</div>
 
 			<div
-				className="ps-break-words ps-mt-12"
+				className={`ps-break-words ps-mt-12 ${contentClassName}`}
 				dangerouslySetInnerHTML={{ __html: content }}
 			/>
 		</article>