+++ packages/wordpress-kit/src/lib/tailwindcssPlugin/utilities/index.ts
@@ -1,4 +1,5 @@
 import { ColorUtilities } from './Colors';
 import { FontsUtilities } from './Fonts';
+import { BaseUtilities } from './Base';
 
-export { ColorUtilities, FontsUtilities };
+export { ColorUtilities, FontsUtilities, BaseUtilities };