+++ starters/gatsby-wordpress-starter/__tests__/snapshotTests/homepage.test.jsx
@@ -1,25 +1,17 @@
-import * as Gatsby from 'gatsby'
 import { render } from '@testing-library/react'
 
 import Index from '../../src/templates/index'
 import data from '../data/allWpPost.json'
 
-const StaticQuery = vi.spyOn(Gatsby, 'StaticQuery')
-
 describe('<Index />', () => {
-	beforeEach(() => {
-		StaticQuery.mockImplementation(() => {
-			return {
-				data,
-			}
-		})
-	})
-
 	it('should render with posts', () => {
 		const { asFragment } = render(
 			<Index
-				data={data}
-				pageContext={{ nextPagePath: null, previousPagePath: null }}
+				pageContext={{
+					nextPagePath: null,
+					previousPagePath: null,
+					posts: data,
+				}}
 			/>,
 		)